Refactor : csr slide 컴포넌트화
