+++ src/components/shared/Carousel.js
@@ -1,0 +1,60 @@
+import classNames from 'classnames/bind';
+import { useState, useRef, useEffect } from 'react';
+import { Swiper, SwiperSlide } from 'swiper/react';
+import { Pagination, Navigation } from 'swiper';
+
+import styles from './Carousel.module.scss';
+import 'swiper/css';
+import 'swiper/css/pagination';
+
+import arrow_left_icon from '../../assets/images/shared/arrow-back.svg';
+import arrow_right_icon from '../../assets/images/shared/arrow-next.svg';
+
+const cx = classNames.bind(styles);
+
+function Carousel({ images, slidesPerView }) {
+  const [pager, setPager] = useState(null);
+  const paginationRef = useRef();
+
+  useEffect(() => {
+    setPager(paginationRef.current.className);
+  }, []);
+
+  return (
+    <div className={cx('carousel')}>
+      <Swiper
+        loop={true}
+        slidesPerView={slidesPerView}
+        pagination={{
+          el: `.${pager}`,
+          clickable: true,
+          bulletActiveClass: cx('bulletsActive'),
+        }}
+        navigation={{
+          nextEl: '.next',
+          prevEl: '.prev',
+        }}
+        modules={[Pagination, Navigation]}>
+        {images.map((image, index) => (
+          <SwiperSlide>
+            <img key={index} src={image} alt="" />
+          </SwiperSlide>
+        ))}
+      </Swiper>
+
+      <div className={cx('controls')}>
+        <div ref={paginationRef} className={cx('pager')}></div>
+        <div className={cx('nav')}>
+          <button className={cx('prev')}>
+            <img src={arrow_left_icon} alt="" />
+          </button>
+          <button className={cx('next')}>
+            <img src={arrow_right_icon} alt="" />
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default Carousel;
