+++ src/components/shared/Carousel.module.scss
@@ -1,0 +1,57 @@
+@import '../../common.scss';
+
+.carousel {
+  line-height: 0;
+
+  img {
+    width: 100%;
+  }
+
+  > .controls {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+
+    > .pager {
+      display: flex;
+      align-items: center;
+      height: 20px;
+
+      > span {
+        display: block;
+        width: 12px;
+        height: 12px;
+        border-right: 50%;
+        background-color: #06542c;
+        opacity: 0.3;
+
+        &.bulletsActive {
+          opacity: 1;
+        }
+      }
+    }
+
+    > .nav {
+      @include flexCenterStyle;
+      > button {
+        @include flexCenterStyle;
+        border: 0;
+        width: 3.125rem;
+        height: 3.125rem;
+        background-color: #06542c;
+        cursor: pointer;
+
+        &.prev {
+        }
+
+        &.next {
+        }
+
+        > img {
+          width: 100%;
+          max-width: 0.8rem;
+        }
+      }
+    }
+  }
+}
