+++ src/pages/csr/share/index.module.scss
@@ -55,62 +55,6 @@
           width: 50%;
         }
       }
-
-      > .slide {
-        line-height: 0;
-
-        img {
-          width: 100%;
-        }
-
-        > .controls {
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-
-          > .pager {
-            display: flex;
-            align-items: center;
-            height: 20px;
-
-            > span {
-              display: block;
-              width: 12px;
-              height: 12px;
-              border-right: 50%;
-              background-color: #06542c;
-              opacity: 0.3;
-
-              &.bulletsActive {
-                opacity: 1;
-              }
-            }
-          }
-
-          > .nav {
-            @include flexCenterStyle;
-            > button {
-              @include flexCenterStyle;
-              border: 0;
-              width: 3.125rem;
-              height: 3.125rem;
-              background-color: #06542c;
-              cursor: pointer;
-
-              &.prev {
-              }
-
-              &.next {
-              }
-
-              > img {
-                width: 100%;
-                max-width: 0.8rem;
-              }
-            }
-          }
-        }
-      }
     }
   }
 }
